Fire emission dies down over time & colliders
Changed the fire particle system emissions rate over time into a curve that causes the fire to gradually lower emission over time simulating a dying fire. The glow now loops and remains when the fire dies out.

A collider and rigidbody was added to the firepit to prevent the player from going through it

The area of light was increased